% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectAnalytics.R
\name{connectAnalytics}
\alias{connectAnalytics}
\title{Run the Shiny Application}
\usage{
connectAnalytics(
  ...,
  host = Sys.getenv("RSTUDIO_CONNECT_SERVER"),
  api_key = Sys.getenv("RSTUDIO_CONNECT_API_KEY"),
  user = NULL,
  switch_user = TRUE,
  favicon = NULL,
  title = "connectAnalytics",
  window_title = NULL,
  header_width = 230
)
}
\arguments{
\item{...}{additional options to be passed to `golem_opts`}

\item{host}{the host server for the RStudio Connect to be connected to. 
Defaults to the environment variable "RSTUDIO_CONNECT_SERVER"}

\item{api_key}{a valid RStudio Connect API key}

\item{user}{the desired user to log in as. If NULL (default) it will default
to the session$user (recommended)}

\item{switch_user}{logical; if TRUE a button appears in the navbar that will 
allow users to switch which connect user analytics data they are looking at. 
This could be useful if your company wants to allow users to see how others are
performing.}

\item{favicon}{path to a favicon icon to be used for your application}

\item{title}{the title of the application to be shown in the dashboard header}

\item{window_title}{the title of the application to be shown in the browser tab. 
If NULL (default) the title value will be used}

\item{header_width}{Set the width of the header for the shiny application. This is
useful in case the title specified is too long for the default width}
}
\description{
Run the Shiny Application
}
