% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_01_server_connect.R
\name{mod_01_server_connect_ui}
\alias{mod_01_server_connect_ui}
\alias{mod_01_server_connect_server}
\title{mod_01_server_connect_ui and mod_01_server_connect_server}
\usage{
mod_01_server_connect_ui(id)

mod_01_server_connect_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{r}{a reactiveValues object}
}
\description{
A shiny Module handling the connection to a RStudio Connect
server if the expected environment variables are not valid
}
\keyword{internal}
