% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_username.R
\name{mod_02_username_ui}
\alias{mod_02_username_ui}
\alias{mod_02_username_server}
\title{mod_02_username_ui and mod_02_username_server}
\usage{
mod_02_username_ui(id)

mod_02_username_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
