% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_content.R
\name{mod_04_content_ui}
\alias{mod_04_content_ui}
\alias{mod_04_content_server}
\title{mod_04_content_ui and mod_04_content_server}
\usage{
mod_04_content_ui(id)

mod_04_content_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
