% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_05_usage.R
\name{mod_05_usage_ui}
\alias{mod_05_usage_ui}
\alias{mod_05_usage_server}
\title{mod_05_usage_ui and mod_05_usage_server}
\usage{
mod_05_usage_ui(id, admin = FALSE)

mod_05_usage_server(input, output, session, r, admin = FALSE)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{r}{a reactiveValues object}
}
\description{
A shiny Module to get the usage for the user defined window 
to show them how much their content is being used, when it is being used and 
who is using it.
}
\keyword{internal}
