% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_06_admin.R
\name{mod_06_admin_ui}
\alias{mod_06_admin_ui}
\alias{mod_06_admin_server}
\title{mod_06_admin_ui and mod_06_admin_server}
\usage{
mod_06_admin_ui(id)

mod_06_admin_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
