% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_05_usage.R
\name{overall_usage_line}
\alias{overall_usage_line}
\title{Create a interactive line graph showing Connect Usage}
\usage{
overall_usage_line(usage_tbl, from, to, username, admin = FALSE)
}
\arguments{
\item{usage_tbl}{A tibble output from \code{overall_usage_tbl}
\code{connectapi::get_usage_static} function}

\item{from}{the start date of the interested query}

\item{to}{the end data of the interested query}

\item{username}{the username to put in the graph title}

\item{admin}{whether this is for the admin section}
}
\value{
a plotly object
}
\description{
create a interactive plotly graph showing the 
daily usage trends for shiny application, static content, and 
overall
}
\details{
The username and admin arguments only impact the plot title
}
