% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_05_usage.R
\name{overall_usage_tbl}
\alias{overall_usage_tbl}
\title{Get a tbl showing overall usage by date}
\usage{
overall_usage_tbl(shiny_usage, static_usage, from, to)
}
\arguments{
\item{shiny_usage}{a shiny_usage dataset from the 
\link{\code{connectapi::get_usage_shiny}} function}

\item{static_usage}{a data from the 
\link{\code{connectapi::get_usage_static}} function}

\item{from}{the start date of the interested query}

\item{to}{the end data of the interested query}
}
\value{
a tbl in long format
}
\description{
Get a tbl from content usage data that shows the total number
of sessions each day for static and shiny applications and overall
}
