% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_05_usage.R
\name{usage_info_join}
\alias{usage_info_join}
\title{Join usage info with content and user info}
\usage{
usage_info_join(usage, content, users)
}
\arguments{
\item{usage}{a usage data frame (shiny or static) with a content_guid and 
user_guid column}

\item{content}{a content dataframe that has information about each of the 
pieces of content on the server. Must have columns: guid, owner_username, 
and title}

\item{users}{a users data frame that has information about the users on the 
server. Must have columns: username, first_name, last_name, and guid}
}
\value{
a single dataset that contains the information about each usage instance
including what content it is, who owns the content and who used it
}
\description{
Join the three datasets (usage, content, and users) to return
a single dataset that contains infomation about each of the apps that are 
used on the Connect system
}
\details{
If the content GUID does in the usage dataset does not match any entries in the 
content dataset than it will be marked as "Removed Content"

If the user GUID does not have a match or if it is NA to begin with than the 
user of the content will be marked as "Anonymous"
}
