% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_userinfo.R
\name{mod_03_userinfo_ui}
\alias{mod_03_userinfo_ui}
\alias{mod_03_userinfo_server}
\title{mod_03_userinfo_ui and mod_03_userinfo_server}
\usage{
mod_03_userinfo_ui(id)

mod_03_userinfo_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
