% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_05_usage.R
\name{mod_05_usage_ui}
\alias{mod_05_usage_ui}
\alias{mod_05_usage_server}
\title{mod_05_usage_ui and mod_05_usage_server}
\usage{
mod_05_usage_ui(id)

mod_05_usage_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
