% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_05_usage.R
\name{overall_usage_line}
\alias{overall_usage_line}
\title{Create a interactive line graph showing Connect Usage}
\usage{
overall_usage_line(usage_tbl, from, to, username, admin = FALSE)
}
\arguments{
\item{from}{the start date of the interested query}

\item{to}{the end data of the interested query}

\item{shiny_usage}{a shiny_usage dataset from the 
\code{connectapi::get_usage_shiny} function}

\item{static_usage}{a data from the 
\code{connectapi::get_usage_static} function}
}
\value{
a plotly object
}
\description{
create a interactive plotly graph showing the 
daily usage trends for shiny application, static content, and 
overall
}
